Rename Plugin Service File
