+++ src/services/tunnel.ts
@@ -7,7 +7,7 @@ import { homedir } from "os";
 import { relative, join } from "path";
 
 import { validateAndParseOpenApiSpec } from "./openapi-service";
-import { PluginService } from "./plugin-service";
+import { PluginService } from "./plugin";
 import { BITTE_CONFIG_ENV_KEY } from "../config/constants";
 import { appendToEnv, removeFromEnv } from "../utils/file-utils";
 import { getSpecUrl } from "../utils/url-utils";
